split scene pkg into smaller independent sub pkgs
