+++ tea/scenes/errors/errors.go
@@ -1,4 +1,4 @@
-package scenes
+package errors
 
 const (
 	ErrorFailedAddMod = "failed to add mod"
