+++ tea/scenes/keys/keys.go
@@ -1,4 +1,4 @@
-package scenes
+package keys
 
 const (
 	KeyControlC = "ctrl+c"
