+++ tea/scenes/mods/mod.go
@@ -1,4 +1,4 @@
-package scenes
+package mods
 
 import (
 	"time"
@@ -9,6 +9,8 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/satisfactorymodding/ficsit-cli/tea/components"
+	"github.com/satisfactorymodding/ficsit-cli/tea/scenes/errors"
+	"github.com/satisfactorymodding/ficsit-cli/tea/scenes/keys"
 	"github.com/satisfactorymodding/ficsit-cli/tea/utils"
 )
 
@@ -69,8 +71,8 @@ func NewModMenu(root components.RootModel, parent tea.Model, mod utils.Mod) tea.
 				Activate: func(msg tea.Msg, currentModel modMenu) (tea.Model, tea.Cmd) {
 					err := root.GetCurrentProfile().AddMod(mod.Reference, ">=0.0.0")
 					if err != nil {
-						log.Error().Err(err).Msg(ErrorFailedAddMod)
-						cmd := currentModel.list.NewStatusMessage(ErrorFailedAddMod)
+						log.Error().Err(err).Msg(errors.ErrorFailedAddMod)
+						cmd := currentModel.list.NewStatusMessage(errors.ErrorFailedAddMod)
 						return currentModel, cmd
 					}
 					return currentModel.parent, nil
@@ -114,7 +116,7 @@ func (m modMenu) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch msg := msg.(type) {
 	case tea.KeyMsg:
 		switch keypress := msg.String(); keypress {
-		case KeyControlC:
+		case keys.KeyControlC:
 			return m, tea.Quit
 		case "q":
 			if m.parent != nil {
@@ -122,7 +124,7 @@ func (m modMenu) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				return m.parent, nil
 			}
 			return m, tea.Quit
-		case KeyEnter:
+		case keys.KeyEnter:
 			i, ok := m.list.SelectedItem().(utils.SimpleItem[modMenu])
 			if ok {
 				if i.Activate != nil {
