+++ tea/scenes/profile/messages.go
@@ -1,4 +1,4 @@
-package scenes
+package profile
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
@@ -14,16 +14,4 @@ type updateProfileNames struct{}
 
 func updateProfileNamesCmd() tea.Msg {
 	return updateProfileNames{}
-}
-
-type updateInstallationList struct{}
-
-func updateInstallationListCmd() tea.Msg {
-	return updateInstallationList{}
-}
-
-type updateInstallationNames struct{}
-
-func updateInstallationNamesCmd() tea.Msg {
-	return updateInstallationNames{}
 }
